pyproject.toml
[tool.poetry]
name = "eko-bot"
version = "0.1.0"
description = ""
authors = ["Iuri Ribeiro <iurics10@gmail.com>"]
readme = "README.md"
packages = [{include = "eko_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "*"
python-dotenv = "*"
llama_index = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
